address most PR comments
